refactory generate repay

+++ ttsd-service/src/main/java/com/tuotiansudai/dto/LoanDto.java
@@ -35,7 +35,7 @@ public class LoanDto {
     /***借款期限***/
     @NotEmpty
     @Pattern(regexp = "^\\d+$")
-    private long periods;
+    private int periods;
 
     /***项目描述（纯文本）***/
     @NotEmpty
@@ -153,11 +153,11 @@ public class LoanDto {
         this.type = type;
     }
 
-    public long getPeriods() {
+    public int getPeriods() {
         return periods;
     }
 
-    public void setPeriods(long periods) {
+    public void setPeriods(int periods) {
         this.periods = periods;
     }
